[FRCV-17] Organize the types into a one main folder
## [FRCV-17] Description
Organized all type files under the main types folder.
