+++ src/types/AI.types.ts
